requirements.txt
numpy
tensorflow
librosa
PyYAML
